added multerRoutes to controllers/index.js; migrated code from server.js to multerRoutes.js
